# swagger.yaml
swagger: '2.0'


schemes: 
  - http
  - https

host: localhost:8085
basePath: /bytebank-api/v1

info:
  version: 1.0.0
  title: ByteBank-Api
  description: Administra a movimentação bancaria do correntista e controla o acesso atraves de CPF e senha.
  termsOfService: http://localhost:8085/bytebank-api/v1/termo-uso
  contact:
    name: Suporte
    url: http://localhost:8085/bytebank-api/v1/suporte
    email: suporte@bytebank-api.com


securityDefinitions:
  BasicAuth: 
    type: basic
    
security: 
  - BasicAuth: []
    
consumes:
  - application/json
produces:
  - application/json
  
tags: 
  - name: cliente
  - name: conta
  - name: extrato
  - name: transacao 
        
paths:

  '/cliente':
    post:
      tags:
        - cliente
      summary: Cadastra Cliente
      description: Cadastra um novo cliente
      operationId: cadastraCliente
      parameters:
        - in: body
          name: cliente_
          schema:
            $ref: '#/definitions/Cliente'
            
      responses:
        201:
          description: Cadastro efetuado com sucesso
          schema:
            $ref: '#/definitions/Cliente'      
        400:
          description: Requisição invalida
        500:
          description: Erro Interno do Servidor 
          
  
  '/conta/{agencia}/{numero}/{digito}/saldo':
    get:
      tags:
        - conta
      summary: Consulta Saldo
      description: Consulta o saldo da conta do cliente
      operationId: consultaSaldo
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: agencia
          type: integer
          format: int32
          required: true
          
        - in: path
          name: numero
          type: integer
          format: int64
          required: true
          
        - in: path
          name: digito
          type: integer
          format: int32
          required: true
          
        - in: header
          name: Authorization
          type: string
          required: true
          
      responses:
        200:
          description: Consulta a saldo efetuada com sucesso
          schema:
            type: number
            format: double
        400:
          description: Requisicao Invalida          
        401:
          description: Requisicao não autorizada
        500:
          description: Erro do servidor 
        

        
  '/transacao':
    post:
      tags:
        - transacao
      summary: Cadastra Transacao na conta
      description: cadastra uma nova transacao ou deposito na conta
      security:
        - BasicAuth: []
      operationId: cadastraTransacao
      parameters:
        - in: header
          type: string
          name: Authorization
          required: true
          
        - in: body
          name: transacao 
          schema:
            $ref: '#/definitions/Transacao'
          required: true
          
        - in: query
          name: tipo
          type: string
          enum: 
            - debito
            - deposito
          required: true
      responses:
        201:
          description: Cadastro efetuado com sucesso
          schema:
            $ref: '#/definitions/Transacao'
        400:
          description: Requisição invalida
        401:
          description: requisicao não autorizada
        500:
          description: Erro Interno do Servidor 
          
          
  '/extrato/{agencia}/{numero}/{digito}':
    get:
      tags:
        - extrato
      summary: Consulta Extrato
      description: Consulta o extrato da conta do cliente
      operationId: consultaExtrato
      security:
        - BasicAuth: []
      parameters:
        - in: path
          name: agencia
          type: integer
          format: int32
          required: true
          
        - in: path
          name: numero
          type: integer
          format: int64
          required: true
          
        - in: path
          name: digito
          type: integer
          format: int32
          required: true
          
        - in: header
          name: Authorization
          type: string
          required: true
          
      responses:
        200:
          description: Consulta a extrato efetuada com sucesso
          schema:
            $ref: '#/definitions/Extrato'
        400:
          description: Requisicao Invalida          
        401:
          description: Requisicao não autorizada
        500:
          description: Erro do servidor 
  
  '/transacao/{codigo}':
    put:
      tags:
        - transacao
      summary: Altera Valor da Transação
      description: Altera o valor da transação existente na conta do cliente
      security:
        - BasicAuth: []
      operationId: alteraValorTransacao
      parameters:
        - in: path
          name: codigo
          type: integer
          format: int64
          required: true
          
        - in: query
          name: valor
          type: number
          format: double
          required: true
          
        - in: header
          name: Authorization
          type: string
          required: true
          
      responses:
        204:
          description: Alteração do valor efetuada com sucesso
        400:
          description: Requisição invalida
        401:
          description: requisicao não autorizada
        500:
          description: Erro Interno do Servidor 
    
    delete:
      tags:
        - transacao
      summary: Exclui a transação 
      description: Exclusão da Transacao existente na conta do cliente
      operationId: excluiTransacao
      security:
        - BasicAuth: []      
      parameters:
        - in: path
          name: codigo
          type: integer
          format: int64
          required: true
        
        - in: header
          name: Authorization
          type: string
          required: true      
      responses:
        204:
          description: Transação Excluida com sucesso
        400:
          description: Requisição invalida
        401:
          description: requisicao não autorizada
        500:
          description: Erro Interno do Servidor 

definitions:
  
  Cliente:
    type: object
    properties:
      id: 
        type: integer
        format: int64
        example: 12345
      titular:
        type: string
        example: Janete Silva
      cpf:
        type: string
        maxLength: 11
        minLength: 11
        example: 12345678910
      senha:
        type: string
        format: password
        example: senha123
      conta:
        type: array
        items: 
          $ref: '#/definitions/Conta'    
  
  Conta:
    type: object
    properties:
      agencia:
        type: integer
        format: int32
        example: 1234
      numero:
        type: integer
        format: int64
        example: 123456
      digito:
        type: integer
        format: int32
        example: 0
      saldo:
        type: number
        format: double
        example: 101.34
      cliente:
        $ref: '#/definitions/Cliente'
      transacoes:
        type: array
        items: 
          $ref: '#/definitions/Transacao'
        
  Transacao:
    type: object
    properties:
      id:
        type: integer
        format: int64
      tipo:
        type: string
        enum:
          - debito
          - deposito
      valor:
        type: number
        format: double
      data:
        type: string
        format: date-time
      conta:
        $ref: '#/definitions/Conta'
        
  Extrato:
    type: object
    properties:
      transacoes:
        type: array
        items:
          $ref: '#/definitions/Transacao'
      